chore(progress): split demo
