+++ packages/nutui-taro-demo/src/exhibition/pages/progress/basic.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="30" />
+</template>
