+++ packages/nutui-taro-demo/src/exhibition/pages/progress/color.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="30" stroke-color="skyblue" stroke-width="15" text-color="green" />
+</template>
