+++ packages/nutui-taro-demo/src/exhibition/pages/progress/content.vue
@@ -1,0 +1,9 @@
+<template>
+  <nut-progress :percentage="60" text-inside>
+    <img
+      src="https://img11.360buyimg.com/imagetools/jfs/t1/137646/13/7132/1648/5f4c748bE43da8ddd/a3f06d51dcae7b60.png"
+      width="30"
+      height="30"
+    />
+  </nut-progress>
+</template>
