+++ packages/nutui-taro-demo/src/exhibition/pages/progress/inside.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="60" text-inside />
+</template>
