+++ src/packages/__VUE/progress/demo/basic.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="30" />
+</template>
