+++ src/packages/__VUE/progress/demo/index.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import Color from './color.vue';
 import Text from './text.vue';
