+++ src/packages/__VUE/progress/demo/inside.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="60" text-inside />
+</template>
