+++ src/packages/__VUE/progress/demo/size.vue
@@ -1,0 +1,5 @@
+<template>
+  <nut-progress :percentage="30" text-inside size="small"> </nut-progress>
+  <nut-progress :percentage="50" text-inside size="base"> </nut-progress>
+  <nut-progress :percentage="70" text-inside size="large"> </nut-progress>
+</template>
