+++ src/packages/__VUE/progress/demo/text.vue
@@ -1,0 +1,3 @@
+<template>
+  <nut-progress :percentage="50" :show-text="false" />
+</template>
