+++ src/packages/__VUE/progress/doc.en-US.md
@@ -16,144 +16,37 @@ app.use(Progress);
 
 ### Basic Usage
 
-:::demo
-
-```vue
-<template>
-  <nut-progress percentage="30" />
-</template>
-```
-
-:::
+> demo: progress basic
 
 ### Custom Style
 
-:::demo
-
-```vue
-<template>
-  <nut-progress percentage="30" stroke-color="rgba(250,44,25,0.47)" stroke-width="20" text-color="red" />
-</template>
-```
-
-:::
+> demo: progress color
 
 ### Don't Show Percentage
 
-:::demo
-
-```vue
-<template>
-  <nut-progress percentage="50" :show-text="false" stroke-height="24" />
-</template>
-```
-
-:::
+> demo: progress text
 
 ### Text Inside
 
-:::demo
-
-```vue
-<template>
-  <nut-progress percentage="60" :text-inside="true" />
-</template>
-```
-
-:::
+> demo: progress inside
 
 ### Custom Content
 
-:::demo
-
-```vue
-<template>
-  <nut-progress percentage="60" :text-inside="true">
-    <img
-      src="https://img11.360buyimg.com/imagetools/jfs/t1/137646/13/7132/1648/5f4c748bE43da8ddd/a3f06d51dcae7b60.png"
-      width="30"
-      height="30"
-    />
-  </nut-progress>
-</template>
-```
-
-:::
+> demo: progress content
 
 ### Custom Size
 
-**small**，**base**，**large** .
-:::demo
+size: `small`, `base`, `large` .
 
-```vue
-<template>
-  <nut-progress percentage="30" :text-inside="true" size="small"> </nut-progress>
-  <nut-progress percentage="50" :text-inside="true" size="base"> </nut-progress>
-  <nut-progress percentage="70" :text-inside="true" size="large"> </nut-progress>
-</template>
-```
-
-:::
+> demo: progress size
 
 ### Status Display
 
-:::demo
-
-```vue
-<template>
-  <nut-progress
-    percentage="30"
-    stroke-color="linear-gradient(270deg, rgba(18,126,255,1) 0%,rgba(32,147,255,1) 32.815625%,rgba(13,242,204,1) 100%)"
-    status="active"
-  />
-  <nut-progress percentage="50" status="icon" />
-  <nut-progress
-    percentage="100"
-    stroke-color="linear-gradient(90deg, rgba(180,236,81,1) 0%,rgba(66,147,33,1) 100%)"
-    stroke-width="15"
-    status="icon"
-  >
-    <template #icon-name>
-      <Issue color="red" width="15px" height="15px"></Issue>
-    </template>
-  </nut-progress>
-</template>
-<script setup>
-import { Issue } from '@nutui/icons-vue';
-</script>
-```
-
-:::
+> demo: progress status
 
 ### Dynamic Change
 
-:::demo
-
-```vue
-<template>
-  <nut-progress :percentage="val" />
-  <nut-button type="default" @click="setReduceVal">reduce</nut-button>
-  <nut-button type="primary" @click="setAddVal">add</nut-button>
-</template>
-<script setup>
-import { ref } from 'vue';
-const val = ref(0);
-const setAddVal = () => {
-  if (val.value >= 100) {
-    return false;
-  }
-  val.value += 10;
-};
-const setReduceVal = () => {
-  if (val.value <= 0) {
-    return false;
-  }
-  val.value -= 10;
-};
-</script>
-```
-
-:::
+> demo: progress change
 
 ## API
